batpipe: Fix parent process detection
Less is called with a filename, and if we basename CMD of less we get
a basename of a filename, not a less executable. I think it's just
a type that basename was piped to cut, CMD should be cut'ed before
processed by basename.

As parent_executable is only used by batpipe, we can just cut CMD and
extract program name.